+++ src/main/java/com/gregtechceu/gtceu/api/mui/drawable/graph/MinorTickFinder.java
@@ -5,5 +5,5 @@ import org.jetbrains.annotations.ApiStatus;
 @ApiStatus.Experimental
 public interface MinorTickFinder {
 
-    float[] find(float min, float max, float[] majorTicks, float[] ticks);
+    double[] find(double min, double max, double[] majorTicks, double[] ticks);
 }
